refactor(cli): remove dead commands — gc, import, community
- Gc: dead stub, replaced by Examined: 5 skill(s)
Transitions (1):
  old-skill            Archived -> Draft    (promotion)
Decayed (read): 1 (confidence recomputed on read; persisted only on promotion)
Archived: 0
- Import: dead code, replaced by
- Community: public marketplace, conflicts with strategy; use
- Team functions extracted to new team_cmd.rs module
